requirement.txt
fastapi
uvicorn[standard]
scapy
faiss-cpu
numpy
sentence-transformers
websockets
docker
pyyaml
requests
sqlalchemy
python-dotenv
psutil
cryptography
bcrypt
python-jose[cryptography]
passlib[bcrypt]
alembic
pydantic-settings